rename AbstractMultiSearcher -> MultiSearcher
